+++ stub/ResultPaginator.php
@@ -6,7 +6,7 @@ use GuzzleHttp\Promise;
 use Aws\Result;
 use GuzzleHttp\Promise\PromiseInterface;
 
-class ResultPaginator
+class ResultPaginator implements \Iterator
 {
     /**
      * @var Result
@@ -26,5 +26,28 @@ class ResultPaginator
     public function each(callable $callback)
     {
         return Promise\promise_for($callback($this->result));
+    }
+
+    public function valid()
+    {
+        return $this->result ? true : false;
+    }
+
+    public function current()
+    {
+        return $this->valid() ? $this->result : false;
+    }
+
+    public function next()
+    {
+        $this->result = null;
+    }
+
+    public function key()
+    {
+    }
+
+    public function rewind()
+    {
     }
 }